pyproject.toml
[project]
name = "faculty-appraisal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "sqlalchemy>=2.0.49",
    "supabase>=2.29.0",
    "uvicorn>=0.46.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
